move `modifyDiffOutput` inside helpers
